feat: Adds functionality to docker compose
Adds functionality to run debug containers for admin and ta. This also adds a new functionality to create rust binary from inside the container this helps operating system like MacOs to build binary that can be mounted in linux based containers.

We also introduced a new network in the docker compose file which removes link from the docker compose file since it is deprecated.

Signed-off-by: Farhaan Bukhsh <farhaan.bukhsh@gmail.com>